Move wteg_feedback_widget to feedback directory
